[FE] Multiple changes across files in order to render images properly
