+++ client/src/components/item-details/ItemDetails.module.css
@@ -26,9 +26,10 @@
 }
 
 .mainImage {
+  padding-top: 10px;
   width: 100%;
-  height: 75%;
-  background-color: #C4DDD4;
+  height: 100%;
+  /* background-color: #C4DDD4; */
   border-radius: 4px;
   box-sizing: border-box;
 }
@@ -67,6 +68,8 @@
 
 .sizes {
   display: flex;
+  /* justify-content: center; */
+
   justify-content: space-between;
 }
