+++ client/src/components/item-details/ItemDetails.tsx
@@ -1,5 +1,5 @@
-// import mock from '../../mock-data/mock.json'
-
+import mock from '../../mock-data/mock.json'
+import { Product } from '../../models/models'
 import styles from './ItemDetails.module.css'
 import { useCartSlice } from '../../zustand/ShoppingCartSlice'
 import { useParams } from 'react-router-dom'
@@ -13,42 +13,38 @@ import { ShoppingCartProductType } from '../../../../global-types/shopping-cart-
 
 
 export default function ItemDetails() {
+
   const { id } = userStore()
   const { visible, setVisibility } = menuStore();
-
-  console.log(visible)
   const storeItems = useProductsSlice((state) => state.storeItems)
   const addItem = useCartSlice((state) => state.addItem)
   const openCart = useCartSlice((state) => state.openCart)
 
-  // URL param
-  const param = useParams()
-  // console.log(param.id)
-useEffect(() => {
- setVisibility(false)
-  }, [id]);
- 
+  useEffect(() => {
+  setVisibility(false)
+    }, [id]);
+  
   // URL query and fetch the DB
   // const data = JSON.parse(JSON.stringify(mock))
-  // const products: Product[] = data.products;
+  // const product: Product[] = data.products[0];
+  
+  const param = useParams()
   let product = storeItems.find(item => String(item.id) === param.id)
-
-  // const [item, setItem] = useState(product)
+  console.log(product)
 
   // handle size selection
   const handleSizeSelection = (e: React.SyntheticEvent) => {
     const size = e.currentTarget.textContent as ProductSize // refactor to make more strict type
-    // console.log('PRODUCT:', {...product, size: size})
     // return {...product, size: size}
     if (size && product) {
       // setItem({...product, size: size})
 
       // change this later to useRef
-      const sizes = document.querySelectorAll('.size')
-      sizes.forEach(s => {
-        if (s.textContent === size) s.classList.add('selected')
-        else s.classList.remove('selected')
-      })
+      // const sizes = document.querySelectorAll('.size')
+      // sizes.forEach(s => {
+      //   if (s.textContent === size) s.classList.add('selected')
+      //   else s.classList.remove('selected')
+      // })
     }
   }
 
@@ -57,21 +53,22 @@ useEffect(() => {
 
       {/* LEFT */}
       <div className={styles.left}>
-        <div className={styles.mainImage}>IMG</div>
+        <img className={styles.mainImage} src={product?.image} alt='item_picture'></img>
+        {/* <img className={styles.mainImage} src={product?.pictureUrl} alt='item_picture'></img> */}
         <div className={styles.sideImages}>
+          {/* <div className={styles.sideImage}></div>
           <div className={styles.sideImage}></div>
-          <div className={styles.sideImage}></div>
-          <div className={styles.sideImage}></div>
+          <div className={styles.sideImage}></div> */}
         </div>
       </div>
 
 
       {/* RIGHT */}
       <div className={styles.right}>
-        <h1 className={styles.name}>{product?.name}</h1>
-        {/* <h4 className={styles.description}>{product?.description}</h4> */}
+        <h1 className={styles.name} onClick={handleSizeSelection}>{product?.name}</h1>
+        <h4 className={styles.price} onClick={handleSizeSelection}>{product?.category}</h4>
         <h4 className={styles.price}>{product?.price}</h4>
-
+        
         <div className={styles.sizes}>
           <div className={styles.size} onClick={handleSizeSelection}>S</div>
           <div className={styles.size} onClick={handleSizeSelection}>M</div>
@@ -87,8 +84,9 @@ useEffect(() => {
                 userId: id,
                 productId: product.id,
               });
+
               // check if response === typeof ShoppingCartItem
-              // console.log('-->', newCartItem)
+              console.log('-->', newCartItem)
               if (newCartItem.productQuantity) {
                 addItem(newCartItem)
                 openCart();
