+++ client/src/components/store-items/StoreItems.tsx
@@ -1,13 +1,14 @@
 // import mock from '../../mock-data/mock.json'
 // import { Product } from '../../models/models'
 import styles from './StoreItems.module.css'
-import StoreItem from '../../components/store-item/StoreItem'
+import StoreItem from '../../components/Item/Item'
 import { useEffect } from 'react'
 import { useProductsSlice } from '../../zustand/ProductSlice'
 import { getStoreProducts } from '../../services/store-products-service'
 import { ProductType } from '../../../../global-types/product'
 
 export default function StoreItems() {
+  
   const storeItems = useProductsSlice((state) => state.storeItems)
   const addProduct = useProductsSlice((state) => state.addProduct)
