+++ client/src/models/models.tsx
@@ -1,16 +1,16 @@
 export type ProductSize = 'S' | 'M' | 'L' | 'XL';
 
 
-export interface Product {
-  id: number
-  name: string
-  category: string
-  size: ProductSize
-  price: number
-  sellerId: string
-  quantity: number
-  pictureUrl: string
-  }
+// export interface Product {
+//   id: number
+//   name: string
+//   category: string
+//   size: string
+//   price: number
+//   sellerId: string
+//   quantity: number
+//   image: string
+//   }
 
 export type User = {
   isLoggedIn: boolean
