+++ global-types/product.ts
@@ -1,9 +1,10 @@
 export type ProductType = {
   id: number;
   name: string;
-  category: string;
-  price: number;
-  sellerId: string | null; // uuid
-  quantity: number;
-  pictureUrl: string;
+  category: string
+  price: number
+  sellerId: string
+  quantity: number
+  pictureUrl: string
+  image: string
 };
